+++ packages/sdks-tests/src/tests/helpers.ts
@@ -41,6 +41,18 @@ export const test = base.extend<TestOptions>({
 });
 test.afterEach(screenshotOnFailure);
 
+export const isSSRFramework = (packageName: PackageName | 'DEFAULT') => {
+  // Easier to list non-ssr than other way around.
+  const isNonSSR =
+    packageName === 'solid' ||
+    packageName === 'react' ||
+    packageName === 'vue2' ||
+    packageName === 'svelte' ||
+    packageName === 'react-native' ||
+    packageName === 'gen1-react';
+  return !isNonSSR;
+};
+
 export const findTextInPage = async ({ page, text }: { page: Page; text: string }) => {
   await page.locator(`text=${text}`).waitFor();
 };