+++ packages/sdks/output/svelte/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.6.4
+
+- Fix issue with styles not rendering properly during SSR.
+
+### 0.6.3
+
+- Fix issue with block styles not updating when editing them.
+
 ### 0.6.2
 
 - No Changes.